+++ solver/edge.go
@@ -2,7 +2,6 @@ package solver
 
 import (
 	"context"
-	"strings"
 	"sync"
 	"time"
 
@@ -34,29 +33,8 @@ func newEdge(ed Edge, op activeOp, index *edgeIndex) *edge {
 		cacheRecords:       map[string]*CacheRecord{},
 		cacheRecordsLoaded: map[string]struct{}{},
 		index:              index,
-		debug:              debugScheduler,
 	}
-	if !e.debug && len(debugSchedulerSteps) > 0 {
-		withParents := strings.HasSuffix(debugSchedulerSteps[0], "^")
-		name := strings.TrimSuffix(debugSchedulerSteps[0], "^")
-		for _, v := range debugSchedulerSteps {
-			if strings.Contains(name, v) {
-				e.debug = true
-				break
-			}
-		}
-		if !e.debug && withParents {
-			for _, vtx := range ed.Vertex.Inputs() {
-				name := strings.TrimSuffix(vtx.Vertex.Name(), "^")
-				for _, v := range debugSchedulerSteps {
-					if strings.Contains(name, v) {
-						e.debug = true
-						break
-					}
-				}
-			}
-		}
-	}
+	e.debug = debugSchedulerCheckEdge(e)
 	return e
 }
 
@@ -203,17 +181,13 @@ func (e *edge) finishIncoming(req pipeSender) {
 	if req.Request().Canceled && err == nil {
 		err = context.Canceled
 	}
-	if e.debug {
-		bklog.G(context.TODO()).Debugf("finishIncoming %s %v %#v desired=%s", e.edge.Vertex.Name(), err, e.edgeState, req.Request().Payload.desiredState)
-	}
+	debugSchedulerFinishIncoming(e, err, req)
 	req.Finalize(&e.edgeState, err)
 }
 
 // updateIncoming updates the current value of incoming pipe request
 func (e *edge) updateIncoming(req pipeSender) {
-	if e.debug {
-		bklog.G(context.TODO()).Debugf("updateIncoming %s %#v desired=%s", e.edge.Vertex.Name(), e.edgeState, req.Request().Payload.desiredState)
-	}
+	debugSchedulerUpdateIncoming(e, req)
 	req.Update(&e.edgeState)
 }
 
@@ -707,9 +681,7 @@ func (e *edge) recalcCurrentState() {
 					}
 					if len(openKeys) == 0 {
 						e.state = edgeStatusCacheSlow
-						if e.debug {
-							bklog.G(context.TODO()).Debugf("upgrade to cache-slow because no open keys")
-						}
+						debugSchedulerUpgradeCacheSlow(e)
 					}
 				}
 			}
@@ -746,9 +718,7 @@ func (e *edge) respondToIncoming(incoming []pipeSender, allPipes []pipeReceiver)
 		allIncomingCanComplete = false
 	}
 
-	if e.debug {
-		bklog.G(context.TODO()).Debugf("status state=%s cancomplete=%v hasouts=%v noPossibleCache=%v depsCacheFast=%v keys=%d cacheRecords=%d", e.state, allIncomingCanComplete, e.hasActiveOutgoing, e.noCacheMatchPossible, e.allDepsCompletedCacheFast, len(e.keys), len(e.cacheRecords))
-	}
+	debugSchedulerRespondToIncomingStatus(e, allIncomingCanComplete)
 
 	if allIncomingCanComplete && e.hasActiveOutgoing {
 		// cancel all current requests
@@ -843,44 +813,15 @@ func (e *edge) createInputRequests(desiredState edgeStatusType, f *pipeFactory, 
 			addNew := true
 			if dep.req != nil && !dep.req.Status().Completed {
 				if dep.req.Request().desiredState != desiredStateDep {
-					if e.debug {
-						bklog.G(context.TODO()).
-							WithField("edge_vertex_name", e.edge.Vertex.Name()).
-							WithField("edge_vertex_digest", e.edge.Vertex.Digest()).
-							WithField("dep_index", dep.index).
-							WithField("dep_req_desired_state", dep.req.Request().desiredState).
-							WithField("dep_desired_state", desiredStateDep).
-							WithField("dep_state", dep.state).
-							Debug("cancel input request")
-					}
+					debugSchedulerCancelInputRequest(e, dep, desiredStateDep)
 					dep.req.Cancel()
 				} else {
-					if e.debug {
-						bklog.G(context.TODO()).
-							WithField("edge_vertex_name", e.edge.Vertex.Name()).
-							WithField("edge_vertex_digest", e.edge.Vertex.Digest()).
-							WithField("dep_index", dep.index).
-							WithField("dep_req_desired_state", dep.req.Request().desiredState).
-							WithField("dep_desired_state", desiredStateDep).
-							WithField("dep_state", dep.state).
-							Debug("skip input request based on existing request")
-					}
+					debugSchedulerSkipInputRequestBasedOnExistingRequest(e, dep, desiredStateDep)
 					addNew = false
 				}
 			}
 			if addNew {
-				if e.debug {
-					bklog.G(context.TODO()).
-						WithField("edge_vertex_name", e.edge.Vertex.Name()).
-						WithField("edge_vertex_digest", e.edge.Vertex.Digest()).
-						WithField("dep_index", dep.index).
-						WithField("dep_desired_state", desiredStateDep).
-						WithField("dep_state", dep.state).
-						WithField("dep_vertex_name", e.edge.Vertex.Inputs()[dep.index].Vertex.Name()).
-						WithField("dep_vertex_digest", e.edge.Vertex.Inputs()[dep.index].Vertex.Digest()).
-						Debug("add input request")
-				}
-
+				debugSchedulerAddInputRequest(e, dep, desiredStateDep)
 				req := f.NewInputRequest(e.edge.Vertex.Inputs()[int(dep.index)], &edgeRequest{
 					currentState: dep.edgeState,
 					desiredState: desiredStateDep,
@@ -890,16 +831,8 @@ func (e *edge) createInputRequests(desiredState edgeStatusType, f *pipeFactory, 
 				dep.req = req
 				addedNew = true
 			}
-		} else if e.debug {
-			bklog.G(context.TODO()).
-				WithField("edge_vertex_name", e.edge.Vertex.Name()).
-				WithField("edge_vertex_digest", e.edge.Vertex.Digest()).
-				WithField("dep_index", dep.index).
-				WithField("dep_desired_state", desiredStateDep).
-				WithField("dep_state", dep.state).
-				WithField("dep_vertex_name", e.edge.Vertex.Inputs()[dep.index].Vertex.Name()).
-				WithField("dep_vertex_digest", e.edge.Vertex.Inputs()[dep.index].Vertex.Digest()).
-				Debug("skip input request based on dep state")
+		} else {
+			debugSchedulerSkipInputRequestBasedOnDepState(e, dep, desiredStateDep)
 		}
 		// initialize function to compute cache key based on dependency result
 		if dep.state == edgeStatusComplete && dep.slowCacheReq == nil && (e.slowCacheFunc(dep) != nil || e.preprocessFunc(dep) != nil) && e.cacheMap != nil {
